# Speed up binomial