--- a/src/sage/symbolic/function.pxd
+++ b/src/sage/symbolic/function.pxd
@@ -12,7 +12,7 @@ cdef class Function(SageObject):
     cdef _register_function(self)
 
 cdef class BuiltinFunction(Function):
-    cdef object _preserved_arg
+    cdef int _preserved_arg  # 0 if none, otherwise in [1.._nargs], see function.pyx
     cdef _is_registered(self)
 
 cdef class GinacFunction(BuiltinFunction):