--- a/src/sage/symbolic/function.pxd
+++ b/src/sage/symbolic/function.pxd
@@ -3,9 +3,9 @@ from sage.structure.sage_object cimport SageObject
 cdef class Function(SageObject):
     cdef unsigned int _serial
     cdef int _nargs
-    cdef object _name
-    cdef object _alt_name
-    cdef object _latex_name
+    cdef str _name
+    cdef str _alt_name
+    cdef str _latex_name
     cdef object _conversions
     cdef object _evalf_params_first
     cdef _is_registered(self)
@@ -16,7 +16,7 @@ cdef class BuiltinFunction(Function):
     cdef _is_registered(self)
 
 cdef class GinacFunction(BuiltinFunction):
-    cdef object _ginac_name
+    cdef str _ginac_name
     cdef _is_registered(self)
     cdef _register_function(self)