As first observed in https://github.com/sagemath/sage/pull/39379, when `n` and `k` are small sage integers, `binomial(n, k)` are much slower than `n.binomial(k)`.

This pull request partially fixes it.

There's also a minor patch where some code is skipped when all arguments are instances of `Element`. It is certainly a more common case for the input to be `Element` than otherwise, so we optimize for it.

Benchmark for that part. With the optimization:

```
sage:         %timeit binomial(20, 10)
5.11 μs ± 114 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
sage:         %timeit binomial(20, 10)
4.97 μs ± 42.5 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
sage:         %timeit binomial(20, 10)
5.07 μs ± 43.7 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
sage:         %timeit binomial(20, 10)
5.06 μs ± 134 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
```

Without (note that the ± is not confidence interval, it is just standard deviation):

```
sage:         %timeit binomial(20, 10)
5.45 μs ± 20.9 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
sage:         %timeit binomial(20, 10)
5.37 μs ± 44.8 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
sage:         %timeit binomial(20, 10)
5.39 μs ± 85.2 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
sage:         %timeit binomial(20, 10)
5.47 μs ± 87.8 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
```

Note that it is still slower than direct method call. At least it's less bad, previously it was ≈ 30μs per call.

```
sage: %timeit 20.binomial(10)
1.07 μs ± 17.3 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
```

I think using not-overriding `_method_arguments` as an indicator of "don't call the method"
(then silently catch `TypeError` and throw it away) is rather bad design,
but this is not in scope here.

### :memo: Checklist

<!-- Put an `x` in all the boxes that apply. -->

- [x] The title is concise and informative.
- [x] The description explains in detail what this PR is about.
- [x] I have linked a relevant issue or discussion.
- [x] I have created tests covering the changes.
- [ ] I have updated the documentation and checked the documentation preview.